requirements.txt
websockets>=11.0.3
asyncio==3.4.3
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
# API相关依赖
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
email-validator>=2.0.0
websocket-client>=1.6.3
requests>=2.31.0
httpx>=0.24.1
jinja2>=3.1.2
schedule>=1.2.0
aiohttp>=3.11.16
